Retry start of neo4j service on Windows
Prunsrv sometimes fails to start the neo4j service due to a timeout.
Until the reason for why this timeout occures a one time retry of
starting should make the integration tests green.

When it fails to start it will be visible in the logs so we shouldn't
forget about this problem.